chatgpt-backend/requirements.txt
tiktoken==0.4.0
tokenizers==0.13.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
faiss-cpu==1.7.4
langchain==0.0.206
langchainplus-sdk==0.0.16
pdf2image==1.16.3
pdfminer.six==20221105
pypdf==3.10.0
PyPDF2==3.0.1
stream-chat==4.8.0
streamlit==1.23.1
streamlit-chat==0.0.2.2
python-dotenv
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
six==1.16.0
Werkzeug==2.2.2